we need a base script that can be called directly to update / install the xen components

+++ scripts/install-xen.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 # a small script to install xen
-
-yum -y install centos-release-xen
-yum -y install xen 
-/usr/bin/grub-bootxen.sh
+if [ ! -e /tmp/installed-xen ]; then
+	yum -y install centos-release-xen
+	yum -y install xen 
+	/usr/bin/grub-bootxen.sh
+	touch /tmp/installed-xen
+fi
